Split shared Supabase types out of a giant supabase.ts
